Close connection on ErrPktSync and ErrPktSyncMul
An `ErrPktSync` or `ErrPktSyncMul` error always means that a packet
header has been read, but since the sequence ID was not correct then the
packet payload has not been read. This results in the connection being
left in a broken state, since any future operations will always result
in a "busy buffer" error. Keeping such connections alive leads to them
being repeatedly returned to the pool in this state, which can in turn
result in a large number of failures due to these "busy buffer" errors.

This commit fixes this problem by simply closing the connection before
returning either `ErrPktSync` or `ErrPktSyncMul`. This ensures that the
connection won't be returned to the pool, preventing it from causing any
further errors.